Moved kernel_scheduler into its own file
